# docs/openapi.yml
components:
  schemas:
...
    EsiScope:
      description: Enumeration of the ESI scopes
      enum:
      - esi-alliances.read_contacts.v1
      - esi-assets.read_assets.v1
      - esi-assets.read_corporation_assets.v1
      - esi-bookmarks.read_character_bookmarks.v1
      - esi-bookmarks.read_corporation_bookmarks.v1
      - esi-calendar.read_calendar_events.v1
      - esi-calendar.respond_calendar_events.v1
      - esi-characters.read_agents_research.v1
      - esi-characters.read_blueprints.v1
      - esi-characters.read_chat_channels.v1
      - esi-characters.read_contacts.v1
      - esi-characters.read_corporation_roles.v1
      - esi-characters.read_fatigue.v1
      - esi-characters.read_fw_stats.v1
      - esi-characters.read_loyalty.v1
      - esi-characters.read_medals.v1
      - esi-characters.read_notifications.v1
      - esi-characters.read_opportunities.v1
      - esi-characters.read_standings.v1
      - esi-characters.read_titles.v1
      - esi-characters.write_contacts.v1
      - esi-characterstats.read.v1
      - esi-clones.read_clones.v1
      - esi-clones.read_implants.v1
      - esi-contracts.read_character_contracts.v1
      - esi-contracts.read_corporation_contracts.v1
      - esi-corporations.read_blueprints.v1
      - esi-corporations.read_contacts.v1
      - esi-corporations.read_container_logs.v1
      - esi-corporations.read_corporation_membership.v1
      - esi-corporations.read_divisions.v1
      - esi-corporations.read_facilities.v1
      - esi-corporations.read_fw_stats.v1
      - esi-corporations.read_medals.v1
      - esi-corporations.read_standings.v1
      - esi-corporations.read_starbases.v1
      - esi-corporations.read_structures.v1
      - esi-corporations.read_titles.v1
      - esi-corporations.track_members.v1
      - esi-fittings.read_fittings.v1
      - esi-fittings.write_fittings.v1
      - esi-fleets.read_fleet.v1
      - esi-fleets.write_fleet.v1
      - esi-industry.read_character_jobs.v1
      - esi-industry.read_character_mining.v1
      - esi-industry.read_corporation_jobs.v1
      - esi-industry.read_corporation_mining.v1
      - esi-killmails.read_corporation_killmails.v1
      - esi-killmails.read_killmails.v1
      - esi-location.read_location.v1
      - esi-location.read_online.v1
      - esi-location.read_ship_type.v1
      - esi-mail.organize_mail.v1
      - esi-mail.read_mail.v1
      - esi-mail.send_mail.v1
      - esi-markets.read_character_orders.v1
      - esi-markets.read_corporation_orders.v1
      - esi-markets.structure_markets.v1
      - esi-planets.manage_planets.v1
      - esi-planets.read_customs_offices.v1
      - esi-search.search_structures.v1
      - esi-skills.read_skillqueue.v1
      - esi-skills.read_skills.v1
      - esi-ui.open_window.v1
      - esi-ui.write_waypoint.v1
      - esi-universe.read_structures.v1
      - esi-wallet.read_character_wallet.v1
      - esi-wallet.read_corporation_wallet.v1
      - esi-wallet.read_corporation_wallets.v1
      - publicData
      title: EsiScope
      type: string
...
    GetCorporationOut:
      description: Corporation data
      properties:
        alliance:
          $ref: '#/components/schemas/sni__api__routers__corporation__GetAllianceShortOut'
        authorized_to_login:
          title: Authorized To Login
          type: boolean
        ceo:
          $ref: '#/components/schemas/GetUserShortOut'
        corporation_id:
          title: Corporation Id
          type: integer
        corporation_name:
          title: Corporation Name
          type: string
        cumulated_mandatory_esi_scopes:
          items:
            $ref: '#/components/schemas/EsiScope'
          title: Cumulated Mandatory Esi Scopes
          type: array
        mandatory_esi_scopes:
          items:
            $ref: '#/components/schemas/EsiScope'
          title: Mandatory Esi Scopes
          type: array
        ticker:
          title: Ticker
          type: string
        updated_on:
          format: date-time
          title: Updated On
          type: string
      required:
      - ceo
      - corporation_id
      - corporation_name
      - cumulated_mandatory_esi_scopes
      - mandatory_esi_scopes
      - ticker
      - updated_on
      title: GetCorporationOut
      type: object
    GetCorporationShortOut:
      description: Short corporation description
      properties:
        corporation_id:
          title: Corporation Id
          type: integer
        corporation_name:
          title: Corporation Name
          type: string
      required:
      - corporation_id
      - corporation_name
      title: GetCorporationShortOut
      type: object
...
    GetTrackingOut:
      description: Represents a corporation tracking response.
      properties:
        invalid_refresh_token:
          default: []
          items:
            $ref: '#/components/schemas/GetUserShortOut'
          title: Invalid Refresh Token
          type: array
        no_refresh_token:
          default: []
          items:
            $ref: '#/components/schemas/GetUserShortOut'
          title: No Refresh Token
          type: array
        valid_refresh_token:
          default: []
          items:
            $ref: '#/components/schemas/GetUserShortOut'
          title: Valid Refresh Token
          type: array
      title: GetTrackingOut
      type: object
...
    GetUserShortOut:
      description: Model for an element of ``GET /user`` response
      properties:
        character_id:
          title: Character Id
          type: integer
        character_name:
          title: Character Name
          type: string
      required:
      - character_id
      - character_name
      title: GetUserShortOut
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    PostCorporationGuestOut:
      description: Model for ``POST /corporation/{corporation_id}/guest`` reponses.
      properties:
        state_code:
          title: State Code
          type: string
      required:
      - state_code
      title: PostCorporationGuestOut
      type: object
...
    PutCorporationIn:
      description: Model for ``PUT /corporation/{corporation_id}`` requests
      properties:
        authorized_to_login:
          title: Authorized To Login
          type: boolean
        mandatory_esi_scopes:
          items:
            $ref: '#/components/schemas/EsiScope'
          title: Mandatory Esi Scopes
          type: array
      title: PutCorporationIn
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            type: string
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
    sni__api__routers__corporation__GetAllianceShortOut:
      description: Short alliance description
      properties:
        alliance_id:
          title: Alliance Id
          type: integer
        alliance_name:
          title: Alliance Name
          type: string
      required:
      - alliance_id
      - alliance_name
      title: GetAllianceShortOut
      type: object
...
info:
  title: FastAPI
  version: 0.1.0
openapi: 3.0.2
paths:
...
  /corporation:
    get:
      description: 'Gets the list of corporations registered in this instance. Requires
        a

        clearance level of 0 or more.'
      operationId: get_corporations_corporation_get
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          title: Authorization
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/GetCorporationShortOut'
                title: Response Get Corporations Corporation Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get the list of corporations
      tags:
      - Corporation management
  /corporation/{corporation_id}:
    get:
      description: 'Get informations about a corporation. Note that this corporation
        must be

        registered on SNI'
      operationId: get_corporation_corporation__corporation_id__get
      parameters:
      - in: path
        name: corporation_id
        required: true
        schema:
          title: Corporation Id
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          title: Authorization
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCorporationOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get informations about a corporation
      tags:
      - Corporation management
    post:
      description: 'Manually fetches a corporation from the ESI. Requires a clearance
        level of

        8 or more.'
      operationId: post_corporation_corporation__corporation_id__post
      parameters:
      - in: path
        name: corporation_id
        required: true
        schema:
          title: Corporation Id
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          title: Authorization
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCorporationOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Manually fetch a corporation from the ESI
      tags:
      - Corporation management
    put:
      description: 'Modify a corporation registered on SNI. Note that it does not
        modify it on

        an ESI level. Requires a clearance level of 2 or more.'
      operationId: put_corporation_corporation__corporation_id__put
      parameters:
      - in: path
        name: corporation_id
        required: true
        schema:
          title: Corporation Id
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          title: Authorization
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutCorporationIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCorporationOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Modify a corporation registered on SNI
      tags:
      - Corporation management
  /corporation/{corporation_id}/guest:
    get:
      description: Returns the list of guests in this corporation.
      operationId: get_corporation_guests_corporation__corporation_id__guest_get
      parameters:
      - in: path
        name: corporation_id
        required: true
        schema:
          title: Corporation Id
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          title: Authorization
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/GetUserShortOut'
                title: Response Get Corporation Guests Corporation  Corporation Id  Guest
                  Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Corporation guests
      tags:
      - Corporation management
    post:
      description: 'Creates a state code for a new guest to this corporation. The
        user then has

        to login with this state code to be considered a guest.'
      operationId: post_corporation_guest_corporation__corporation_id__guest_post
      parameters:
      - in: path
        name: corporation_id
        required: true
        schema:
          title: Corporation Id
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          title: Authorization
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostCorporationGuestOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Creates a state code for a new guest to this corporation
      tags:
      - Corporation management
  /corporation/{corporation_id}/guest/{character_id}:
    delete:
      description: Deletes a corporation guest
      operationId: delete_corporation_guest_corporation__corporation_id__guest__character_id__delete
      parameters:
      - in: path
        name: corporation_id
        required: true
        schema:
          title: Corporation Id
          type: integer
      - in: path
        name: character_id
        required: true
        schema:
          title: Character Id
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          title: Authorization
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Deletes a corporation guest
      tags:
      - Corporation management
  /corporation/{corporation_id}/tracking:
    get:
      description: 'Reports which member (of a given corporation) have a valid refresh
        token

        attacked to them, and which do not. Requires a clearance level of 1 and

        having authority over this corporation.'
      operationId: get_corporation_tracking_corporation__corporation_id__tracking_get
      parameters:
      - in: path
        name: corporation_id
        required: true
        schema:
          title: Corporation Id
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          title: Authorization
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTrackingOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Corporation tracking
      tags:
      - Corporation management
...